+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 - [Button & ItemChoice] API breaking changes, now use a unique `status` instead of multiple props to avoid overlaps of properties and styles
 - [Radio & AutoComplete] Repercussions from ItemChoice API changes
 
+#v0.1.0
+- [Transitions] Add Transitions component
+- [Modal] Apply Transitions component and add unit tests
+
 #v0.0.2
 - [Button] Shadowed prop on secondary button